+++ Swift/testprog/enums/enums.structures.swift
@@ -64,7 +64,7 @@ extension EnumsHeader where T: BP3DProto.Buffer {
     }
     public var type: EnumsType? {
         let rawValue = self.rawType;
-        return EnumsType(rawValue: UInt(rawValue));
+        return EnumsType(rawValue: rawValue);
     }
 
 }
